% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_duration.R
\name{calc_duration}
\alias{calc_duration}
\title{Calculate the duration of a code for a datavyu column}
\usage{
calc_duration(column = NULL, code = NULL, directory = NULL, by_file = FALSE)
}
\arguments{
\item{column}{the column as a character string}

\item{code}{the code as a character string}

\item{directory}{the path to the directory as a character string}

\item{by_file}{whether or not to calculate the frequencies by file (logical)}
}
\value{
A data frame generated with the janitor package
}
\description{
Calculate the duration of a code for a datavyu column
}
\examples{
calc_frequencies(column = "LogClass_AS_ActivityFormat",
                 directory = "ex-data/datavyu_output_07-06-2020_14-46",
                 by_file = TRUE)

}
