% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_frequencies.R
\name{calc_frequencies}
\alias{calc_frequencies}
\title{Calculate code frequencies for a datavu column}
\usage{
calc_frequencies(column = NULL, code = NULL, directory = NULL, by_file = FALSE)
}
\arguments{
\item{column}{the column as a character string}

\item{code}{the code as a character string; defaults to code01}

\item{directory}{the path to the directory as a character string}

\item{by_file}{whether or not to calculate the frequencies by file (logical)}
}
\value{
A data frame generated with the janitor package
}
\description{
Calculate code frequencies for a datavu column
}
