% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-find-unique-files.R
\name{find_unique_files}
\alias{find_unique_files}
\title{Find unique columns in a directory}
\usage{
find_unique_files(directory = NULL)
}
\arguments{
\item{directory}{the path to the directory as a character string; can be set with options(directory = "")}
}
\value{
A vector with the unique columns in a directory
}
\description{
Find unique columns in a directory
}
\examples{
find_unique_files("ex-data/datavyu_output_07-06-2020_14-46")
}
