% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-methods.R
\name{plot.datavyu}
\alias{plot.datavyu}
\title{Plot datavyu output}
\usage{
\method{plot}{datavyu}(datavyu_object)
}
\arguments{
\item{datavyu_object}{a datavyu object created by \code{summarize_column()}}
}
\description{
Plot datavyu output
}
