% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-duration.R
\name{plot_duration}
\alias{plot_duration}
\title{Plot code duration for a datavu column}
\usage{
plot_duration(datavyu_object)
}
\arguments{
\item{datavyu_object}{a datavyu object created by \code{summarize_column()}}
}
\value{
A ggplot2 plot
}
\description{
Plot code duration for a datavu column
}
