% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-durations.R
\name{plot_durations}
\alias{plot_durations}
\title{Calculate code durations for a datavu column}
\usage{
plot_durations(column = NULL, code = NULL, folder = NULL, by_file = FALSE)
}
\arguments{
\item{column}{the column as a character string}

\item{code}{the code as a character string}

\item{folder}{the path to the folder as a character string}

\item{by_file}{whether or not to calculate the frequencies by file (logical)}
}
\value{
A data frame generated with the janitor package
}
\description{
Calculate code durations for a datavu column
}
\examples{

}
