% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-frequencies.R
\name{plot_frequencies}
\alias{plot_frequencies}
\title{Plot code frequencies for a datavu column}
\usage{
plot_frequencies(datavyu_object)
}
\arguments{
\item{column}{the column as a character string}

\item{code}{the code as a character string}

\item{folder}{the path to the folder as a character string}

\item{by_file}{whether or not to calculate the frequencies by file (logical)}
}
\value{
A data frame generated with the janitor package
}
\description{
Plot code frequencies for a datavu column
}
\examples{

}
