% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-plot-time-series.R
\name{plot_time_series}
\alias{plot_time_series}
\title{Plot code frequency for a datavyu column}
\usage{
plot_time_series(
  columns,
  directory = NULL,
  colors = 1,
  cat_order = NULL,
  color_scale = NULL
)
}
\arguments{
\item{columns}{the column or columns to select as character string or vector}

\item{directory}{path to where csv files are to be found with \code{datavyur::import_datavyu()}}

\item{colors}{color palette to use, options from \code{ggplot2::scale_color_brewer()} function}

\item{cat_order}{character vector of the order in which to display codes in plot}

\item{color_scale}{a named character vector manually specifying a color scale will also reorder based on order provided}
}
\value{
A ggplot2 plot
}
\description{
Plot code frequency for a datavyu column
}
