% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-plot-time-series.R
\name{plot_time_series_NEW}
\alias{plot_time_series_NEW}
\title{Plot code frequency for a datavyu column IN A NEW WAY}
\usage{
plot_time_series_NEW(datafile)
}
\arguments{
\item{datafile}{the file with onsets and offsets}

\item{normalize_ts}{whether or not to normalize the time stamps to start at 0; defaults to FALSE}
}
\value{
A ggplot2 plot
}
\description{
Plot code frequency for a datavyu column IN A NEW WAY
}
\examples{
\dontrun{
prepared_time_series <- prep_time_series(column = "childhands", code = "childhands.hand",
  specified_file = "dyad1", directory = "ex-data/datavyu_output_11-16-2020_13-26")
  plot_time_series_NEW(prepared_time_series)
}
}
