% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-methods.R
\name{print.datavyu}
\alias{print.datavyu}
\title{Print datavyu output}
\usage{
\method{print}{datavyu}(datavyu_object)
}
\arguments{
\item{datavyu_object}{a datavyu object created by \code{summarize_column()}}
}
\description{
Print datavyu output
}
