% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-methods.R
\name{summary.datavyu}
\alias{summary.datavyu}
\title{Summarize datavyu output}
\usage{
\method{summary}{datavyu}(datavyu_object)
}
\arguments{
\item{datavyu_object}{a datavyu object created by \code{summarize_column()}}
}
\description{
Summarize datavyu output
}
