% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-column-summary.R
\name{plot_column_summary}
\alias{plot_column_summary}
\title{Plot code duration for a datavu column}
\usage{
plot_column_summary(datavyu_object)
}
\arguments{
\item{datavyu_object}{a datavyu object created by \code{summarize_column()}}
}
\value{
A ggplot2 plot
}
\description{
Plot code duration for a datavu column
}
\examples{

freq_summary <- summarize_column(column = "LogClass_AS_ActivityFormat",
                                 directory = "ex-data/datavyu_output_07-06-2020_14-46")

freq_summary_by_file <- summarize_column(column = "LogClass_AS_ActivityFormat",
                                         directory = "ex-data/datavyu_output_07-06-2020_14-46",
                                         by_file = TRUE)

plot_column_summary(freq_summary)
plot_column_summary(freq_summary_by_file)

duration_summary <- summarize_column(column = "LogClass_AS_ActivityFormat",
                                     directory = "ex-data/datavyu_output_07-06-2020_14-46",
                                     what = "duration")

duration_summary_by_file <- summarize_column(column = "LogClass_AS_ActivityFormat",
                                             directory = "ex-data/datavyu_output_07-06-2020_14-46",
                                             what = "duration",
                                             by_file = TRUE)

plot_column_summary(duration_summary)
plot_column_summary(duration_summary_by_file)

}
