% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-duration.R
\name{plot_duration}
\alias{plot_duration}
\title{Plot code duration for a datavu column}
\usage{
plot_duration(datavyu_object)
}
\arguments{
\item{datavyu_object}{a datavyu object created by \code{summarize_column()}}
}
\value{
A ggplot2 plot
}
\description{
Plot code duration for a datavu column
}
\examples{
duration_summary <- summarize_column(column = "LogClass_AS_ActivityFormat",
                                     directory = "ex-data/datavyu_output_07-06-2020_14-46",
                                     summary = "duration")

duration_summary_by_file <- summarize_column(column = "LogClass_AS_ActivityFormat",
                                             directory = "ex-data/datavyu_output_07-06-2020_14-46",
                                             summary = "duration",
                                             by_file = TRUE)

plot_duration(freq_summary)
plot_duration(freq_summary_by_file)

}
