% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-frequency.R
\name{plot_frequency}
\alias{plot_frequency}
\title{Plot code frequency for a datavu column}
\usage{
plot_frequency(datavyu_object)
}
\arguments{
\item{datavyu_object}{a datavyu object created by \code{summarize_column()}}
}
\value{
A ggplot2 plot
}
\description{
Plot code frequency for a datavu column
}
\examples{

freq_summary <- summarize_column(column = "LogClass_AS_ActivityFormat",
                                 directory = "ex-data/datavyu_output_07-06-2020_14-46")

freq_summary_by_file <- summarize_column(column = "LogClass_AS_ActivityFormat",
                                         directory = "ex-data/datavyu_output_07-06-2020_14-46",
                                         by_file = TRUE)

plot_frequency(freq_summary)
plot_frequency(freq_summary_by_file)
}
