% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-plot-time-series.R
\name{plot_time_series}
\alias{plot_time_series}
\title{Plot code frequency for a datavu column}
\usage{
plot_time_series(datavyu_ts_object)
}
\arguments{
\item{data_ts_object}{the output from prep_time_series()}
}
\value{
A ggplot2 plot
}
\description{
Plot code frequency for a datavu column
}
\examples{
prepared_time_series <- prep_time_series(column = "LogClass_AS_ActivityFormat",
                                         file = "MM T102 14-02-17 Content Log",
                                         directory = "ex-data/datavyu_output_07-06-2020_14-46")

plot_time_series(prepared_time_series)

}
