% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-plot-time-series.R
\name{plot_time_series}
\alias{plot_time_series}
\title{Plot code frequency for a datavu column}
\usage{
plot_time_series(datavyu_ts_object, normalize_ts = FALSE)
}
\arguments{
\item{datavyu_ts_object}{the output from prep_time_series()}

\item{normalize_ts}{whether or not to normalize the time stamps to start at 0; defaults to FALSE}
}
\value{
A ggplot2 plot
}
\description{
Plot code frequency for a datavu column
}
\examples{
\dontrun{
prepared_time_series <- prep_time_series(column = "childhands", code = "childhands.hand",
  specified_file = "dyad1", directory = "ex-data/datavyu_output_11-16-2020_13-26")
  plot_time_series(prepared_time_series)
}
}
