% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-time-series.R
\name{plot_time_series}
\alias{plot_time_series}
\title{Plot code frequency for a datavu column}
\usage{
plot_time_series(column, code = NULL, directory, by_file = FALSE)
}
\arguments{
\item{column}{the column as a character string}

\item{code}{the code as a character string; defaults to code01}

\item{directory}{the path to the directory as a character string}

\item{by_file}{whether or not to calculate the frequencies by file (logical)}

\item{what}{either "frequency" or "duration"}
}
\value{
A data frame generated with the janitor package
}
\description{
Plot code frequency for a datavu column
}
\examples{
summarize_column(column = "LogClass_AS_ActivityFormat",
                 directory = "example-data/datavyu_output_07-06-2020_14-46")

}
