% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-prep-time-series.R
\name{prep_time_series}
\alias{prep_time_series}
\title{Plot code frequency for a datavu column}
\usage{
prep_time_series(
  column,
  code,
  specified_file = NULL,
  directory = NULL,
  units = "s"
)
}
\arguments{
\item{column}{the column as a character string}

\item{code}{the code as a character string}

\item{specified_file}{the file name; use datavyur::datavyu_col_search() to determine their names}

\item{directory}{the path to the directory as a character string; can be set with options(directory = "")}

\item{units}{the units the data will be prepared in; either "m" (minutes), "s" (default; seconds), or "ms" (milli-seconds)}
}
\value{
a data frame
}
\description{
Plot code frequency for a datavu column
}
\examples{

}
