% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-summarize-column.R
\name{summarize_column}
\alias{summarize_column}
\title{Summarize a datavyu column}
\usage{
summarize_column(
  column = NULL,
  code = NULL,
  directory = NULL,
  by_file = FALSE,
  summary = "frequency"
)
}
\arguments{
\item{column}{the column as a character string}

\item{code}{the code as a character string; defaults to code01}

\item{directory}{the path to the directory as a character string; can be set with options(directory = "")}

\item{by_file}{whether or not to calculate the frequencies by file (logical)}

\item{summary}{either "frequency" or "duration"}
}
\value{
A data frame generated with the janitor package
}
\description{
Summarize a datavyu column
}
\examples{
summarize_column(column = "LogClass_AS_ActivityFormat",
                 directory = "ex-data/datavyu_output_07-06-2020_14-46")

summarize_column(column = "LogClass_AS_ActivityFormat",
                 directory = "ex-data/datavyu_output_07-06-2020_14-46",
                 by_file = TRUE)

summarize_column(column = "LogClass_AS_ActivityFormat",
                 directory = "ex-data/datavyu_output_07-06-2020_14-46",
                 by_file = TRUE,
                 summary = "duration")
}
